chore: align better with style and structure
